asdasdasd

+++ mainWindow.cs
@@ -17,6 +17,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Collections;
 
+
 namespace ALL_LEGIT
 {
     public partial class MainWindow : Form
@@ -31,6 +32,10 @@ namespace ALL_LEGIT
 
         private async void MainWindow_Load(object sender, EventArgs e)
         {
+
+            RemDL.Checked = Properties.Settings.Default.RemDL;
+            RemCP.Checked = Properties.Settings.Default.RemCP;
+
             if (Properties.Settings.Default.DownloadDir == null || Properties.Settings.Default.DownloadDir.Length < 3)
             {
                 Properties.Settings.Default.DownloadDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.UserProfile)}\\Downloads\\AllLegit Downloads";
@@ -102,7 +107,7 @@ namespace ALL_LEGIT
                 {
                     obj = getJson(CheckURL);
                     Activated = bool.Parse(obj.data.activated.ToString());
-                    await Task.Delay(100);
+                    await Task.Delay(10000);
                 }
 
                 if (Activated)
@@ -126,14 +131,25 @@ namespace ALL_LEGIT
 
 
         }
-        private void downloadFiles(string URL, string FILENAME)
+        private async Task downloadFiles(string URL, string FILENAME, string MagnetNAME)
         {
-            string DL = Properties.Settings.Default.DownloadDir + "\\" + FILENAME;
+            string DIR = Properties.Settings.Default.DownloadDir + "\\" + MagnetNAME;
+            if (!Directory.Exists(DIR))
+            {
+                Directory.CreateDirectory(DIR);
+            }
+            string DL = Properties.Settings.Default.DownloadDir + "\\" + MagnetNAME + FILENAME;
             if (File.Exists(DL))
             {
-
+                DialogResult Overwrite1 = MessageBox.Show("File found, do you want to overwrite?", "Overwrite?", MessageBoxButtons.YesNo);
+                if (Overwrite1 != DialogResult.Yes)
+                {
                     File.Delete(DL);
-                
+                }
+                else
+                {
+                    return;
+                }
             }
             WebClient webClient = new WebClient();
             webClient.DownloadProgressChanged += (s, e) =>
@@ -142,16 +158,19 @@ namespace ALL_LEGIT
             };
             webClient.DownloadFileCompleted += (s, e) =>
             {
-                dlProg.Visible = false;
-                foreach (ListViewItem item in listView1.Items)
+                if (RemDL.Checked)
                 {
-                if (item.SubItems[1].Text.Equals(URL))
+                    foreach (ListViewItem item in listView1.Items)
                     {
-                        listView1.Items.Remove(item);
+                        if (item.SubItems[1].Text.Equals(URL))
+                        {
+                            listView1.Items.Remove(item);
+                        }
                     }
                 }
+                dlProg.Value = 0;
             };
-            webClient.DownloadFileAsync(new Uri(URL),
+            await webClient.DownloadFileTaskAsync(new Uri(URL),
                 $"{DL}");
 
 
@@ -165,45 +184,91 @@ namespace ALL_LEGIT
                 {
                     var obj = getJson($"magnet/upload?agent={apiNAME}&apikey={APIKEY}&magnets[]={pasted}");
                     string magnetID = obj.data.magnets[0].id.ToString();
+                    string magnetName = obj.data.magnets[0].name.ToString();
                     obj = getJson($"magnet/status?agent={apiNAME}&apikey={APIKEY}&id={magnetID}");
                     foreach (var key in obj.data.magnets.links)
                     {
-         
+                        bool skip = false;
                         var result = getJson($"link/unlock?agent={apiNAME}&apikey={APIKEY}&link={key.link}");
                         string unlockedLink = result.data.link.ToString();
-                        listView1.Items.Add(new ListViewItem(new string[] {key.filename.ToString(), unlockedLink}));
-                    }
+                        foreach (ListViewItem item in listView1.Items)
+                        {
+                            if (item.SubItems[0].Text.Equals(key.filename.ToString()) && item.SubItems[2].Text.Equals(magnetName))
+                            {
+                                skip = true;
+                            }
+                        }
+                        if (!skip)
+                        {
+                            listView1.Items.Add(new ListViewItem(new string[] { key.filename.ToString(), unlockedLink, magnetName }));
+
+                        }
+                        foreach (ListViewItem item in listView1.Items)
+                        {
+                            if (item.SubItems[2].Text.Equals(magnetName))
+                            {
+                                item.Checked = true;
+
+                            }
+                        }
+                    }              
                     listView1.Update();
-                    foreach (ListViewItem item in listView1.Items)
+                    dlProg.Value = 0;
+
+                }
+                if (pasted.ToLower().StartsWith("https://"))
+                {
+                    string[] pastedsplit = pasted.Split('\n');
+                    foreach (string s in pastedsplit)
                     {
-                        item.Checked = true;
+                        s.Trim();
+                        var obj = getJson($"link/unlock?agent={apiNAME}&apikey={APIKEY}&link={s}");
+                        string unlockedLink = obj.data.link.ToString();
+                        listView1.Items.Add(new ListViewItem(new string[] {obj.data.filename.ToString(), unlockedLink, obj.data.host.ToString() }));
+                        foreach (ListViewItem item in listView1.Items)
+                        {
+                            if (item.SubItems[0].Text.Equals(obj.data.filename.ToString()) && item.SubItems[1].Text.Equals(unlockedLink))
+                            {
+                                item.Checked = true;
+                            }
+                        }
+
                     }
-                    listView1.Update();
 
                 }
             }
                 return base.ProcessCmdKey(ref msg, keyData);
         }
 
-        private void startDownloads_Click(object sender, EventArgs e)
+        public static bool isDownloading = false;
+        private async void startDownloads_Click(object sender, EventArgs e)
         {
-            if (listView1.CheckedItems.Count > 0)
+            if (isDownloading)
             {
-                foreach (ListViewItem item in listView1.Items)
-                {
-                    dlProg.Value = 0;
-           
-                    if (item.Checked)
-                    {
-                        DownloadingText.Text = $"Downloading {item.SubItems[0].Text}...";
-                        downloadFiles(item.SubItems[1].Text, item.SubItems[0].Text);
-                    }
-
-                }
+                MessageBox.Show("Please allow current download to finish before trying to start more!");
+                return;
             }
             else
             {
-                MessageBox.Show("Please select items to download or hit clear.");
+                isDownloading = true;
+                if (listView1.CheckedItems.Count > 0)
+                {
+                    foreach (ListViewItem item in listView1.Items)
+                    {
+                        if (item.Checked)
+                        {
+                            DownloadingText.Text = $"Downloading {item.SubItems[0].Text}...";
+                            await downloadFiles(item.SubItems[1].Text, item.SubItems[0].Text, item.SubItems[2].Text);
+                        }
+                    }
+                    dlProg.Visible = false;
+                    DownloadingText.Text = "";
+                }
+                else
+                {
+                    MessageBox.Show("Please select items to download or hit clear.");
+                }
+                isDownloading = false;
             }
         }
 
@@ -216,9 +281,22 @@ namespace ALL_LEGIT
                 {
                     if (item.Checked)
                     {
-                        var result = getJson($"link/unlock?agent={apiNAME}&apikey={APIKEY}&link={item.SubItems[1].Text}");
-                        forclip += result.data.link.ToString() + "\n";
+                        forclip += item.SubItems[1].Text + "\n";
+                        if (RemCP.Checked)
+                        {
+                            listView1.Items.Remove(item);
+                        }
                     }
+
+                }
+                forclip = forclip.Trim('\r', '\n');
+                try
+                {
+                    Clipboard.SetText(forclip);
+                }
+                catch
+                {
+                    MessageBox.Show("Clipboard was in use and could not be set!");
                 }
             }
             else
@@ -244,6 +322,19 @@ namespace ALL_LEGIT
                 Properties.Settings.Default.Save();
             }
             DownloadDir.Text = Properties.Settings.Default.DownloadDir;
+        }
+
+        private void RemDL_CheckedChanged(object sender, EventArgs e)
+        {
+            Properties.Settings.Default.RemDL = RemDL.Checked;
+            Properties.Settings.Default.Save();
+        }
+
+        private void RemCP_CheckedChanged(object sender, EventArgs e)
+        {
+            Properties.Settings.Default.RemCP = RemCP.Checked;
+            Properties.Settings.Default.Save();
+
         }
     }
 }
